+++ .phpstorm.meta.php
@@ -2,6 +2,9 @@
 namespace PHPSTORM_META {
 
     $STATIC_METHOD_TYPES = [
+        \Psr\Container\ContainerInterface::get('') => [
+            "" == "@",
+        ],
         \Interop\Container\ContainerInterface::get('') => [
             "" == "@",
         ],