+++ src/DI/Definition/Resolver/ObjectCreator.php
@@ -10,8 +10,8 @@ use DI\Definition\ObjectDefinition\PropertyInjection;
 use DI\DependencyException;
 use DI\Proxy\ProxyFactory;
 use Exception;
-use Interop\Container\Exception\NotFoundException;
 use ProxyManager\Proxy\LazyLoadingInterface;
+use Psr\Container\NotFoundExceptionInterface;
 use ReflectionClass;
 use ReflectionProperty;
 
@@ -141,7 +141,7 @@ class ObjectCreator implements DefinitionResolver
             }
 
             $this->injectMethodsAndProperties($object, $definition);
-        } catch (NotFoundException $e) {
+        } catch (NotFoundExceptionInterface $e) {
             throw new DependencyException(sprintf(
                 'Error while injecting dependencies into %s: %s',
                 $classReflection->getName(),