+++ tests/UnitTest/Definition/StringDefinitionTest.php
@@ -7,7 +7,7 @@ use DI\Definition\StringDefinition;
 use DI\NotFoundException;
 use DI\Scope;
 use EasyMock\EasyMock;
-use Interop\Container\ContainerInterface;
+use Psr\Container\ContainerInterface;
 
 /**
  * @covers \DI\Definition\StringDefinition