+++ change-log.md
@@ -8,6 +8,7 @@ Improvements:
 
 BC breaks:
 
+- PHP 7 or greater is required
 - The deprecated `DI\link()` helper was removed, used `DI\get()` instead
 
 ## 5.4